D% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBIB.R
\name{Dis_BIB}
\alias{Dis_BIB}
\title{Diseño de bloques incompletos balanceados}
\usage{
Dis_BIB(tratamientos, bloques)
}
\arguments{
\item{bloques}{(vector) bloques}

\item{tratamientos(vector)}{tratamientos}
}
\value{
un bloxplot con
}
\description{
Realiza un calculo de anova para un diseño de bloques incompletos
}
